+++ src/components/MessageSimple/MessageContent.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Dimensions, Text } from 'react-native';
+import { Dimensions } from 'react-native';
 import moment from 'moment';
 import { MessageContentContext } from '../../context';
 import styled from '@stream-io/styled-components';
@@ -86,6 +86,43 @@ const FailedText = styled.Text`
   margin-right: 5px;
 `;
 
+const ActionSheetTitleContainer = styled.View`
+  width: 100%;
+  height: 100%;
+  align-items: center;
+  justify-content: center;
+  ${({ theme }) => theme.message.actionSheet.titleContainer.css};
+`;
+
+const ActionSheetTitleText = styled.Text`
+  color: #757575;
+  font-size: 14;
+  ${({ theme }) => theme.message.actionSheet.titleText.css};
+`;
+
+const ActionSheetButtonContainer = styled.View`
+  height: 50;
+  align-items: center;
+  background-color: #fff;
+  justify-content: center;
+  ${({ theme }) => theme.message.actionSheet.buttonContainer.css};
+`;
+
+const ActionSheetButtonText = styled.Text`
+  font-size: 18;
+  color: #388cea;
+  ${({ theme }) => theme.message.actionSheet.buttonText.css};
+`;
+
+const ActionSheetCancelButtonContainer = styled.View`
+  ${({ theme }) => theme.message.actionSheet.cancelButtonContainer.css};
+`;
+const ActionSheetCancelButtonText = styled.Text`
+  font-size: 18;
+  color: red;
+  ${({ theme }) => theme.message.actionSheet.cancelButtonText.css};
+`;
+
 export const MessageContent = themed(
   class MessageContent extends React.PureComponent {
     static themePath = 'message.content';
@@ -104,8 +141,12 @@ export const MessageContent = themed(
        * */
       onThreadSelect: PropTypes.func,
       /**
-       * Callback handler for onPress event on message component
-       */
+       * Callback for onPress event on Message component
+       *
+       * @param e       Event object for onPress event
+       * @param message Message object which was pressed
+       *
+       * */
       onMessageTouch: PropTypes.func,
       /**
        * Handler to delete a current message.
@@ -300,6 +341,7 @@ export const MessageContent = themed(
       const contentProps = {
         alignment: pos,
         status: message.status,
+        onPress: this.props.onMessageTouch,
         onLongPress: options.length > 1 ? this.showActionSheet : null,
         activeOpacity: 0.7,
         disabled: readOnly,
@@ -422,8 +464,29 @@ export const MessageContent = themed(
               ref={(o) => {
                 this.ActionSheet = o;
               }}
-              title={<Text>Choose an action</Text>}
-              options={options.map((o) => o.title)}
+              title={
+                <ActionSheetTitleContainer>
+                  <ActionSheetTitleText>Choose an action</ActionSheetTitleText>
+                </ActionSheetTitleContainer>
+              }
+              options={[
+                ...options.map((o, i) => {
+                  if (i === 0) {
+                    return (
+                      <ActionSheetCancelButtonContainer>
+                        <ActionSheetCancelButtonText>
+                          Cancel
+                        </ActionSheetCancelButtonText>
+                      </ActionSheetCancelButtonContainer>
+                    );
+                  }
+                  return (
+                    <ActionSheetButtonContainer key={o.title}>
+                      <ActionSheetButtonText>{o.title}</ActionSheetButtonText>
+                    </ActionSheetButtonContainer>
+                  );
+                }),
+              ]}
               cancelButtonIndex={0}
               destructiveButtonIndex={0}
               onPress={(index) => this.onActionPress(options[index].id)}